-prepare resip/stack for merge to main  - windows build fixes  - added missing license blocks  - svn eol-style: native  - removed duplicate ConsumerFifoBuffer.hxx (also in rutil)  - added missing files to VS2010 projects  - all projects in main solution now build and link under windows -fixed bug in TuSelector with uninitialized mCongestionManager member

+++ resip/stack/TransactionController.cxx
@@ -168,7 +168,7 @@ TransactionController::send(SipMessage* msg)
    {
       // Need to 503 this.
       SipMessage* resp(Helper::makeResponse(*msg, 503));
-      resp->header(h_RetryAfter).value()=mStateMacFifo.expectedWaitTimeMilliSec()/1000;
+      resp->header(h_RetryAfter).value()=(UInt32)mStateMacFifo.expectedWaitTimeMilliSec()/1000;
       resp->setTransactionUser(msg->getTransactionUser());
       mTuSelector.add(resp, TimeLimitFifo<Message>::InternalElement);
       delete msg;
