+++ resip/stack/Transport.hxx
@@ -1,6 +1,8 @@
 #if !defined(RESIP_TRANSPORT_HXX)
 #define RESIP_TRANSPORT_HXX
 
+#include <stdint.h>
+
 #include "rutil/BaseException.hxx"
 #include "rutil/Data.hxx"
 #include "rutil/FdSetIOObserver.hxx"
@@ -338,7 +340,7 @@ class Transport : public FdSetIOObserver
 
       uint32_t getExpectedWaitForIncoming() const
       {
-         return mStateMachineFifo.getFifo().expectedWaitTimeMilliSec()/1000;
+         return (uint32_t)mStateMachineFifo.getFifo().expectedWaitTimeMilliSec()/1000;
       }
 
       // called by Connection to deliver a received message
